versao de tests true e false...

+++ src/main/java/DAO/LivroDao.java
@@ -3,6 +3,7 @@ package DAO;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import com.mysql.jdbc.Connection;
@@ -40,7 +41,7 @@ public  class LivroDao implements Acervo<Livro> {
 			 return true;
 		}catch(SQLException e){
 
-			logger.error(e.getMessage());
+			logger.error("erro na inserção dos dados"+e.getMessage());
 			
 		}
 		sql.close();
@@ -52,19 +53,18 @@ public  class LivroDao implements Acervo<Livro> {
 
 	public boolean editar(String titulo_livro,String troca_titulo)throws SQLException {
 		PreparedStatement sql =null;
+		LivroDao ld = new LivroDao();
 		try {
+			if(ld.pesquisar(titulo_livro)==true) {
 			sql = conexion.prepareStatement("UPDATE livro SET titulo = ?  WHERE titulo = ?");
 			sql.setString(1, troca_titulo);
 			sql.setString(2, titulo_livro);
 			sql.execute();
-			
-			sql.close();
-			conexion.close();
-			
 			return true;
+			}
 		} catch (SQLException e) {
-			logger.error("falha ao editar!");
-			e.printStackTrace();
+			logger.error("falha ao editar!"+e.getMessage());
+			
 		}
 		sql.close();
 		conexion.close();
@@ -76,39 +76,46 @@ public  class LivroDao implements Acervo<Livro> {
 
 	public boolean pesquisar(String titulo_livro) throws SQLException {
 		PreparedStatement sql =null;
+		
 		try {
 			sql = conexion.prepareStatement("SELECT * FROM livro WHERE titulo = ?");
 			sql.setString(1, titulo_livro);
-			sql.execute();
+			ResultSet result =sql.executeQuery();
+			
+
+			
+			return result.next();
+		} catch (SQLException e) {
+			logger.error("falha na pesquisa!"+ e.getMessage());
+			
+		}finally {
 			sql.close();
 			conexion.close();
-			
-			return true;
-		} catch (SQLException e) {
-			logger.error("falha na pesquisa!");
-			e.printStackTrace();
 		}
-		sql.close();
-		conexion.close();
+		
 		
 		return false;
 	}
 
 
 	public boolean excluir(String titulo_livro) throws SQLException {
+		LivroDao ld = new LivroDao();
 		PreparedStatement sql =null;
 		try {
+			if(ld.pesquisar(titulo_livro) == true) {
 			sql = conexion.prepareStatement("DELETE FROM livro WHERE titulo = ?");
 			sql.setString(1, titulo_livro);
-			sql.execute();
+			sql.executeUpdate();
 			
-			sql.close();
-			conexion.close();
+			boolean result = ld.pesquisar(titulo_livro);
+			if(result == false) {
+				return true;
+			}
+			}
 			
-			return true;
 		} catch (SQLException e) {
-			logger.error("falha!");
-			e.printStackTrace();
+			logger.error("falha!"+e.getMessage());
+			
 		}
 		sql.close();
 		conexion.close();
